fix: general file cleanup
