+++ Creative/apps/creative-tv/src/utils/config.ts
@@ -1,6 +1,5 @@
-import { ThemingProps } from '@chakra-ui/react'
+import { ThemingProps, extendTheme } from '@chakra-ui/react'
 import { Goerli, Mumbai, Polygon } from '@thirdweb-dev/chains'
-import { extendTheme } from '@chakra-ui/react'
 import gql from 'graphql-tag'
 
 export const SITE_NAME = 'CREATIVE TV'
@@ -25,6 +24,7 @@ export const SITE_LOGO = '/grant-logo.png'
   }
 } */
 export const NEXT_PUBLIC_STUDIO_API_KEY = process.env.NEXT_PUBLIC_STUDIO_API_KEY
+export const NEXT_PUBLIC_THIRDWEB_API_KEY = process.env.NEXT_PUBLIC_THIRDWEB_API_KEY
 export const SESSION_PASSWORD = process.env.NEXT_SESSION_PASSWORD
 export const NODE_ENV = process.env.NODE_ENV
 export const SIWE_SESSION_SECRET = process.env.NEXT_PUBLIC_SIWE_SESSION_SECRET
@@ -74,17 +74,17 @@ export const CAROUSEL_PLAYLISTS = {
     description: 'The most popular content on Creative TV.',
     name_1: 'The Wizards Hat ⌐◨-◨ Animated Short Film',
     playbackId_1: 'd82af4v0nip5tka7',
-    name_2:'DAO Palace Documentary',
+    name_2: 'DAO Palace Documentary',
     playbackId_2: 'a2b2fcgqx7ghfpxj',
     name_3: 'This is Nouns',
     playbackId_3: '4d8b5ud89j96qkwt',
     name_4: 'Episode 1 - Wait a minute, who are you?',
     playbackId_4: '6461dsqs9qjr1dji',
-  }
+  },
 }
 
 // BLOCKCHAIN LIST
-export const ETH_CHAINS = [ Goerli]
+export const ETH_CHAINS = [Goerli]
 export const POLYGON_CHAINS = [Polygon]
 export const MUMBAI_CHAIN = [Mumbai]
 
@@ -128,7 +128,7 @@ export const ACCOUNT_FACTORY_TESTNET = '0xEA69aD73011E39E223873214a1F98aFF008A8a
 export const ACCOUNT_FACTORY_MUMBAI = ''
 
 // Setup the Smart Wallet configuration
-export const SMART_WALLET_CONFIG= {
+export const SMART_WALLET_CONFIG = {
   chain: Goerli, // the chain where your smart wallet will be or is deployed
   factoryAddress: ACCOUNT_FACTORY_TESTNET, // your own deployed account factory address
   thirdwebApiKey: SMART_WALLET_KEY, // obtained from the thirdweb dashboard
@@ -156,7 +156,7 @@ export const UNLOCK_QUERY_HOLDS_KEY = gql`
 // Livepeer API
 export const LIVEPEER_API_URL = 'https://livepeer.com/api'
 export const LIVEPEER_HERO_PLAYBACK_ID = 'cbd1dw72qst9xmps' // Welcome To Creative Organization DAO
-export const LIVEPEER_FEATURED_PLAYBACK_ID = '5c2bzf537qbq0r7o'// The Creative Podcast Episode 03
+export const LIVEPEER_FEATURED_PLAYBACK_ID = '5c2bzf537qbq0r7o' // The Creative Podcast Episode 03
 
 export const SERVER_SESSION_SETTINGS = {
   cookieName: SITE_NAME,
@@ -315,9 +315,7 @@ export const FOUNDRY_FACET_ABI = [
       { internalType: 'address', name: 'recipient', type: 'address' },
     ],
     name: 'burn',
-    outputs: [
-      { internalType: 'uint256', name: 'assetsReturned', type: 'uint256' },
-    ],
+    outputs: [{ internalType: 'uint256', name: 'assetsReturned', type: 'uint256' }],
     stateMutability: 'nonpayable',
     type: 'function',
   },
@@ -328,9 +326,7 @@ export const FOUNDRY_FACET_ABI = [
       { internalType: 'address', name: 'sender', type: 'address' },
     ],
     name: 'calculateAssetsReturned',
-    outputs: [
-      { internalType: 'uint256', name: 'assetsReturned', type: 'uint256' },
-    ],
+    outputs: [{ internalType: 'uint256', name: 'assetsReturned', type: 'uint256' }],
     stateMutability: 'view',
     type: 'function',
   },
@@ -340,9 +336,7 @@ export const FOUNDRY_FACET_ABI = [
       { internalType: 'uint256', name: 'assetsDeposited', type: 'uint256' },
     ],
     name: 'calculateMeTokensMinted',
-    outputs: [
-      { internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' },
-    ],
+    outputs: [{ internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' }],
     stateMutability: 'view',
     type: 'function',
   },
@@ -363,9 +357,7 @@ export const FOUNDRY_FACET_ABI = [
       { internalType: 'address', name: 'recipient', type: 'address' },
     ],
     name: 'mint',
-    outputs: [
-      { internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' },
-    ],
+    outputs: [{ internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' }],
     stateMutability: 'nonpayable',
     type: 'function',
   },
@@ -380,9 +372,7 @@ export const FOUNDRY_FACET_ABI = [
       { internalType: 'bytes32', name: 'sSig', type: 'bytes32' },
     ],
     name: 'mintWithPermit',
-    outputs: [
-      { internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' },
-    ],
+    outputs: [{ internalType: 'uint256', name: 'meTokensMinted', type: 'uint256' }],
     stateMutability: 'nonpayable',
     type: 'function',
   },
@@ -391,25 +381,19 @@ export const FOUNDRY_FACET_ABI = [
 export const HUB_FACET_ABI = [
   {
     anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' },
-    ],
+    inputs: [{ indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' }],
     name: 'CancelUpdate',
     type: 'event',
   },
   {
     anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' },
-    ],
+    inputs: [{ indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' }],
     name: 'Deactivate',
     type: 'event',
   },
   {
     anonymous: false,
-    inputs: [
-      { indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' },
-    ],
+    inputs: [{ indexed: false, internalType: 'uint256', name: 'id', type: 'uint256' }],
     name: 'FinishUpdate',
     type: 'event',
   },
@@ -1091,2028 +1075,2026 @@ export const METOKENS_REGISTRY_ABI = [
 ]
 
 export const METOKENS_DIAMOND_ABI = [
-
   {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  }
-  ],
-  "name": "getCurveInfo",
-  "outputs": [
-  {
-  "components": [
-  {
-  "internalType": "uint256",
-  "name": "baseY",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetBaseY",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint32",
-  "name": "reserveWeight",
-  "type": "uint32"
-  },
-  {
-  "internalType": "uint32",
-  "name": "targetReserveWeight",
-  "type": "uint32"
-  }
-  ],
-  "internalType": "struct LibCurve.CurveInfo",
-  "name": "",
-  "type": "tuple"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "meTokensBurned",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "supply",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  }
-  ],
-  "name": "viewAssetsReturned",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "supply",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  }
-  ],
-  "name": "viewMeTokensMinted",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "meTokensBurned",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "supply",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  }
-  ],
-  "name": "viewTargetAssetsReturned",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "supply",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  }
-  ],
-  "name": "viewTargetMeTokensMinted",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "components": [
-  {
-  "internalType": "bytes4[]",
-  "name": "functionSelectors",
-  "type": "bytes4[]"
-  },
-  {
-  "internalType": "address",
-  "name": "facetAddress",
-  "type": "address"
-  },
-  {
-  "internalType": "enum IDiamondCut.FacetCutAction",
-  "name": "action",
-  "type": "uint8"
-  }
-  ],
-  "indexed": false,
-  "internalType": "struct IDiamondCut.FacetCut[]",
-  "name": "diamondCut",
-  "type": "tuple[]"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "init",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "bytes",
-  "name": "data",
-  "type": "bytes"
-  }
-  ],
-  "name": "DiamondCut",
-  "type": "event"
-  },
-  {
-  "inputs": [
-  {
-  "components": [
-  {
-  "internalType": "bytes4[]",
-  "name": "functionSelectors",
-  "type": "bytes4[]"
-  },
-  {
-  "internalType": "address",
-  "name": "facetAddress",
-  "type": "address"
-  },
-  {
-  "internalType": "enum IDiamondCut.FacetCutAction",
-  "name": "action",
-  "type": "uint8"
-  }
-  ],
-  "internalType": "struct IDiamondCut.FacetCut[]",
-  "name": "cut",
-  "type": "tuple[]"
-  },
-  {
-  "internalType": "address",
-  "name": "init",
-  "type": "address"
-  },
-  {
-  "internalType": "bytes",
-  "name": "data",
-  "type": "bytes"
-  }
-  ],
-  "name": "diamondCut",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "bytes4",
-  "name": "functionSelector",
-  "type": "bytes4"
-  }
-  ],
-  "name": "facetAddress",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "facetAddress_",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "facetAddresses",
-  "outputs": [
-  {
-  "internalType": "address[]",
-  "name": "facetAddresses_",
-  "type": "address[]"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "facet",
-  "type": "address"
-  }
-  ],
-  "name": "facetFunctionSelectors",
-  "outputs": [
-  {
-  "internalType": "bytes4[]",
-  "name": "facetFunctionSelectors_",
-  "type": "bytes4[]"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "facets",
-  "outputs": [
-  {
-  "components": [
-  {
-  "internalType": "address",
-  "name": "facetAddress",
-  "type": "address"
-  },
-  {
-  "internalType": "bytes4[]",
-  "name": "functionSelectors",
-  "type": "bytes4[]"
-  }
-  ],
-  "internalType": "struct IDiamondLoupeFacet.Facet[]",
-  "name": "facets_",
-  "type": "tuple[]"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "bytes4",
-  "name": "interfaceId",
-  "type": "bytes4"
-  }
-  ],
-  "name": "supportsInterface",
-  "outputs": [
-  {
-  "internalType": "bool",
-  "name": "",
-  "type": "bool"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "SetBurnBuyerFee",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "SetBurnOwnerFee",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "SetMintFee",
-  "type": "event"
-  },
-  {
-  "inputs": [],
-  "name": "burnBuyerFee",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "burnOwnerFee",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "mintFee",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "setBurnBuyerFee",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "setBurnOwnerFee",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "rate",
-  "type": "uint256"
-  }
-  ],
-  "name": "setMintFee",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "burner",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "recipient",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "meTokensBurned",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "assetsReturned",
-  "type": "uint256"
-  }
-  ],
-  "name": "Burn",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "donor",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  }
-  ],
-  "name": "Donate",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "depositor",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "recipient",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "meTokensMinted",
-  "type": "uint256"
-  }
-  ],
-  "name": "Mint",
-  "type": "event"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "meTokensBurned",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "recipient",
-  "type": "address"
-  }
-  ],
-  "name": "burn",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "meTokensBurned",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "sender",
-  "type": "address"
-  }
-  ],
-  "name": "calculateAssetsReturned",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "assetsReturned",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  }
-  ],
-  "name": "calculateMeTokensMinted",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "meTokensMinted",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  }
-  ],
-  "name": "donate",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "recipient",
-  "type": "address"
-  }
-  ],
-  "name": "mint",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "recipient",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "deadline",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint8",
-  "name": "vSig",
-  "type": "uint8"
-  },
-  {
-  "internalType": "bytes32",
-  "name": "rSig",
-  "type": "bytes32"
-  },
-  {
-  "internalType": "bytes32",
-  "name": "sSig",
-  "type": "bytes32"
-  }
-  ],
-  "name": "mintWithPermit",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "CancelUpdate",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "Deactivate",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "FinishUpdate",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "targetRefundRatio",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint32",
-  "name": "targetReserveWeight",
-  "type": "uint32"
-  },
-  {
-  "indexed": false,
-  "internalType": "bool",
-  "name": "reconfigure",
-  "type": "bool"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "startTime",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "endTime",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "endCooldown",
-  "type": "uint256"
-  }
-  ],
-  "name": "InitUpdate",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "vault",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "refundRatio",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "baseY",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint32",
-  "name": "reserveWeight",
-  "type": "uint32"
-  },
-  {
-  "indexed": false,
-  "internalType": "bytes",
-  "name": "encodedVaultArgs",
-  "type": "bytes"
-  }
-  ],
-  "name": "Register",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "newOwner",
-  "type": "address"
-  }
-  ],
-  "name": "TransferHubOwnership",
-  "type": "event"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "cancelUpdate",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "count",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "deactivate",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "finishUpdate",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  }
-  ],
-  "name": "getHubInfo",
-  "outputs": [
-  {
-  "components": [
-  {
-  "internalType": "uint256",
-  "name": "startTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endCooldown",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "refundRatio",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetRefundRatio",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "internalType": "address",
-  "name": "vault",
-  "type": "address"
-  },
-  {
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "internalType": "bool",
-  "name": "updating",
-  "type": "bool"
-  },
-  {
-  "internalType": "bool",
-  "name": "reconfigure",
-  "type": "bool"
-  },
-  {
-  "internalType": "bool",
-  "name": "active",
-  "type": "bool"
-  }
-  ],
-  "internalType": "struct HubInfo",
-  "name": "",
-  "type": "tuple"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "hubCooldown",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "hubDuration",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "hubWarmup",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetRefundRatio",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint32",
-  "name": "targetReserveWeight",
-  "type": "uint32"
-  }
-  ],
-  "name": "initUpdate",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "internalType": "contract IVault",
-  "name": "vault",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "refundRatio",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "baseY",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint32",
-  "name": "reserveWeight",
-  "type": "uint32"
-  },
-  {
-  "internalType": "bytes",
-  "name": "encodedVaultArgs",
-  "type": "bytes"
-  }
-  ],
-  "name": "register",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "cooldown",
-  "type": "uint256"
-  }
-  ],
-  "name": "setHubCooldown",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "duration",
-  "type": "uint256"
-  }
-  ],
-  "name": "setHubDuration",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "warmup",
-  "type": "uint256"
-  }
-  ],
-  "name": "setHubWarmup",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "id",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "newOwner",
-  "type": "address"
-  }
-  ],
-  "name": "transferHubOwnership",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": true,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "CancelResubscribe",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "from",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "CancelTransferMeTokenOwnership",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "from",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "to",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "ClaimMeTokenOwnership",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": true,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "FinishResubscribe",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": true,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "targetHubId",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "migration",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "bytes",
-  "name": "encodedMigrationArgs",
-  "type": "bytes"
-  }
-  ],
-  "name": "InitResubscribe",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": true,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": true,
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "minted",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "asset",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  },
-  {
-  "indexed": false,
-  "internalType": "string",
-  "name": "name",
-  "type": "string"
-  },
-  {
-  "indexed": false,
-  "internalType": "string",
-  "name": "symbol",
-  "type": "string"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  }
-  ],
-  "name": "Subscribe",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "from",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "to",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "TransferMeTokenOwnership",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "bool",
-  "name": "add",
-  "type": "bool"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "amount",
-  "type": "uint256"
-  }
-  ],
-  "name": "UpdateBalanceLocked",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "bool",
-  "name": "add",
-  "type": "bool"
-  },
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "amount",
-  "type": "uint256"
-  }
-  ],
-  "name": "UpdateBalancePooled",
-  "type": "event"
-  },
-  {
-  "anonymous": false,
-  "inputs": [
-  {
-  "indexed": false,
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "indexed": false,
-  "internalType": "uint256",
-  "name": "newBalance",
-  "type": "uint256"
-  }
-  ],
-  "name": "UpdateBalances",
-  "type": "event"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "cancelResubscribe",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "cancelTransferMeTokenOwnership",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "oldOwner",
-  "type": "address"
-  }
-  ],
-  "name": "claimMeTokenOwnership",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "finishResubscribe",
-  "outputs": [
-  {
-  "components": [
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balanceLocked",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "startTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endCooldown",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetHubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "migration",
-  "type": "address"
-  }
-  ],
-  "internalType": "struct MeTokenInfo",
-  "name": "",
-  "type": "tuple"
-  }
-  ],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  }
-  ],
-  "name": "getMeTokenInfo",
-  "outputs": [
-  {
-  "components": [
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balancePooled",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "balanceLocked",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "startTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endTime",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "endCooldown",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetHubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "migration",
-  "type": "address"
-  }
-  ],
-  "internalType": "struct MeTokenInfo",
-  "name": "",
-  "type": "tuple"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  }
-  ],
-  "name": "getOwnerMeToken",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "oldOwner",
-  "type": "address"
-  }
-  ],
-  "name": "getPendingOwner",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "targetHubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "address",
-  "name": "migration",
-  "type": "address"
-  },
-  {
-  "internalType": "bytes",
-  "name": "encodedMigrationArgs",
-  "type": "bytes"
-  }
-  ],
-  "name": "initResubscribe",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "owner",
-  "type": "address"
-  }
-  ],
-  "name": "isOwner",
-  "outputs": [
-  {
-  "internalType": "bool",
-  "name": "",
-  "type": "bool"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "meTokenCooldown",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "meTokenDuration",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "meTokenWarmup",
-  "outputs": [
-  {
-  "internalType": "uint256",
-  "name": "",
-  "type": "uint256"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "cooldown",
-  "type": "uint256"
-  }
-  ],
-  "name": "setMeTokenCooldown",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "duration",
-  "type": "uint256"
-  }
-  ],
-  "name": "setMeTokenDuration",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "uint256",
-  "name": "warmup",
-  "type": "uint256"
-  }
-  ],
-  "name": "setMeTokenWarmup",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "string",
-  "name": "name",
-  "type": "string"
-  },
-  {
-  "internalType": "string",
-  "name": "symbol",
-  "type": "string"
-  },
-  {
-  "internalType": "uint256",
-  "name": "hubId",
-  "type": "uint256"
-  },
-  {
-  "internalType": "uint256",
-  "name": "assetsDeposited",
-  "type": "uint256"
-  }
-  ],
-  "name": "subscribe",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newOwner",
-  "type": "address"
-  }
-  ],
-  "name": "transferMeTokenOwnership",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "meToken",
-  "type": "address"
-  },
-  {
-  "internalType": "uint256",
-  "name": "newBalance",
-  "type": "uint256"
-  }
-  ],
-  "name": "updateBalances",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "deactivateController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "diamondController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "durationsController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "feesController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "meTokenRegistryController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "registerController",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setDeactivateController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setDiamondController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+    ],
+    name: 'getCurveInfo',
+    outputs: [
+      {
+        components: [
+          {
+            internalType: 'uint256',
+            name: 'baseY',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'targetBaseY',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint32',
+            name: 'reserveWeight',
+            type: 'uint32',
+          },
+          {
+            internalType: 'uint32',
+            name: 'targetReserveWeight',
+            type: 'uint32',
+          },
+        ],
+        internalType: 'struct LibCurve.CurveInfo',
+        name: '',
+        type: 'tuple',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'meTokensBurned',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'supply',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'balancePooled',
+        type: 'uint256',
+      },
+    ],
+    name: 'viewAssetsReturned',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'supply',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'balancePooled',
+        type: 'uint256',
+      },
+    ],
+    name: 'viewMeTokensMinted',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'meTokensBurned',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'supply',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'balancePooled',
+        type: 'uint256',
+      },
+    ],
+    name: 'viewTargetAssetsReturned',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'supply',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'balancePooled',
+        type: 'uint256',
+      },
+    ],
+    name: 'viewTargetMeTokensMinted',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        components: [
+          {
+            internalType: 'bytes4[]',
+            name: 'functionSelectors',
+            type: 'bytes4[]',
+          },
+          {
+            internalType: 'address',
+            name: 'facetAddress',
+            type: 'address',
+          },
+          {
+            internalType: 'enum IDiamondCut.FacetCutAction',
+            name: 'action',
+            type: 'uint8',
+          },
+        ],
+        indexed: false,
+        internalType: 'struct IDiamondCut.FacetCut[]',
+        name: 'diamondCut',
+        type: 'tuple[]',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'init',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'bytes',
+        name: 'data',
+        type: 'bytes',
+      },
+    ],
+    name: 'DiamondCut',
+    type: 'event',
+  },
+  {
+    inputs: [
+      {
+        components: [
+          {
+            internalType: 'bytes4[]',
+            name: 'functionSelectors',
+            type: 'bytes4[]',
+          },
+          {
+            internalType: 'address',
+            name: 'facetAddress',
+            type: 'address',
+          },
+          {
+            internalType: 'enum IDiamondCut.FacetCutAction',
+            name: 'action',
+            type: 'uint8',
+          },
+        ],
+        internalType: 'struct IDiamondCut.FacetCut[]',
+        name: 'cut',
+        type: 'tuple[]',
+      },
+      {
+        internalType: 'address',
+        name: 'init',
+        type: 'address',
+      },
+      {
+        internalType: 'bytes',
+        name: 'data',
+        type: 'bytes',
+      },
+    ],
+    name: 'diamondCut',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'bytes4',
+        name: 'functionSelector',
+        type: 'bytes4',
+      },
+    ],
+    name: 'facetAddress',
+    outputs: [
+      {
+        internalType: 'address',
+        name: 'facetAddress_',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'facetAddresses',
+    outputs: [
+      {
+        internalType: 'address[]',
+        name: 'facetAddresses_',
+        type: 'address[]',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'facet',
+        type: 'address',
+      },
+    ],
+    name: 'facetFunctionSelectors',
+    outputs: [
+      {
+        internalType: 'bytes4[]',
+        name: 'facetFunctionSelectors_',
+        type: 'bytes4[]',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'facets',
+    outputs: [
+      {
+        components: [
+          {
+            internalType: 'address',
+            name: 'facetAddress',
+            type: 'address',
+          },
+          {
+            internalType: 'bytes4[]',
+            name: 'functionSelectors',
+            type: 'bytes4[]',
+          },
+        ],
+        internalType: 'struct IDiamondLoupeFacet.Facet[]',
+        name: 'facets_',
+        type: 'tuple[]',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'bytes4',
+        name: 'interfaceId',
+        type: 'bytes4',
+      },
+    ],
+    name: 'supportsInterface',
+    outputs: [
+      {
+        internalType: 'bool',
+        name: '',
+        type: 'bool',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'SetBurnBuyerFee',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'SetBurnOwnerFee',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'SetMintFee',
+    type: 'event',
+  },
+  {
+    inputs: [],
+    name: 'burnBuyerFee',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'burnOwnerFee',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'mintFee',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'setBurnBuyerFee',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'setBurnOwnerFee',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'rate',
+        type: 'uint256',
+      },
+    ],
+    name: 'setMintFee',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'burner',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'recipient',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'meTokensBurned',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'assetsReturned',
+        type: 'uint256',
+      },
+    ],
+    name: 'Burn',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'donor',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+    ],
+    name: 'Donate',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'depositor',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'recipient',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'meTokensMinted',
+        type: 'uint256',
+      },
+    ],
+    name: 'Mint',
+    type: 'event',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'meTokensBurned',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'recipient',
+        type: 'address',
+      },
+    ],
+    name: 'burn',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'meTokensBurned',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'sender',
+        type: 'address',
+      },
+    ],
+    name: 'calculateAssetsReturned',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: 'assetsReturned',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+    ],
+    name: 'calculateMeTokensMinted',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: 'meTokensMinted',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+    ],
+    name: 'donate',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'recipient',
+        type: 'address',
+      },
+    ],
+    name: 'mint',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'recipient',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'deadline',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint8',
+        name: 'vSig',
+        type: 'uint8',
+      },
+      {
+        internalType: 'bytes32',
+        name: 'rSig',
+        type: 'bytes32',
+      },
+      {
+        internalType: 'bytes32',
+        name: 'sSig',
+        type: 'bytes32',
+      },
+    ],
+    name: 'mintWithPermit',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'CancelUpdate',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'Deactivate',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'FinishUpdate',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'targetRefundRatio',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint32',
+        name: 'targetReserveWeight',
+        type: 'uint32',
+      },
+      {
+        indexed: false,
+        internalType: 'bool',
+        name: 'reconfigure',
+        type: 'bool',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'startTime',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'endTime',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'endCooldown',
+        type: 'uint256',
+      },
+    ],
+    name: 'InitUpdate',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'vault',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'refundRatio',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'baseY',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'uint32',
+        name: 'reserveWeight',
+        type: 'uint32',
+      },
+      {
+        indexed: false,
+        internalType: 'bytes',
+        name: 'encodedVaultArgs',
+        type: 'bytes',
+      },
+    ],
+    name: 'Register',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
+    ],
+    name: 'TransferHubOwnership',
+    type: 'event',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'cancelUpdate',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'count',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'deactivate',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'finishUpdate',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+    ],
+    name: 'getHubInfo',
+    outputs: [
+      {
+        components: [
+          {
+            internalType: 'uint256',
+            name: 'startTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endCooldown',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'refundRatio',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'targetRefundRatio',
+            type: 'uint256',
+          },
+          {
+            internalType: 'address',
+            name: 'owner',
+            type: 'address',
+          },
+          {
+            internalType: 'address',
+            name: 'vault',
+            type: 'address',
+          },
+          {
+            internalType: 'address',
+            name: 'asset',
+            type: 'address',
+          },
+          {
+            internalType: 'bool',
+            name: 'updating',
+            type: 'bool',
+          },
+          {
+            internalType: 'bool',
+            name: 'reconfigure',
+            type: 'bool',
+          },
+          {
+            internalType: 'bool',
+            name: 'active',
+            type: 'bool',
+          },
+        ],
+        internalType: 'struct HubInfo',
+        name: '',
+        type: 'tuple',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'hubCooldown',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'hubDuration',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'hubWarmup',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'targetRefundRatio',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint32',
+        name: 'targetReserveWeight',
+        type: 'uint32',
+      },
+    ],
+    name: 'initUpdate',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+      {
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        internalType: 'contract IVault',
+        name: 'vault',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'refundRatio',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'baseY',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint32',
+        name: 'reserveWeight',
+        type: 'uint32',
+      },
+      {
+        internalType: 'bytes',
+        name: 'encodedVaultArgs',
+        type: 'bytes',
+      },
+    ],
+    name: 'register',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'cooldown',
+        type: 'uint256',
+      },
+    ],
+    name: 'setHubCooldown',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'duration',
+        type: 'uint256',
+      },
+    ],
+    name: 'setHubDuration',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'warmup',
+        type: 'uint256',
+      },
+    ],
+    name: 'setHubWarmup',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'id',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
+    ],
+    name: 'transferHubOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'CancelResubscribe',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'from',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'CancelTransferMeTokenOwnership',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'from',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'to',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'ClaimMeTokenOwnership',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'FinishResubscribe',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'targetHubId',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'migration',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'bytes',
+        name: 'encodedMigrationArgs',
+        type: 'bytes',
+      },
+    ],
+    name: 'InitResubscribe',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'minted',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'asset',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+      {
+        indexed: false,
+        internalType: 'string',
+        name: 'name',
+        type: 'string',
+      },
+      {
+        indexed: false,
+        internalType: 'string',
+        name: 'symbol',
+        type: 'string',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+    ],
+    name: 'Subscribe',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'from',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'to',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'TransferMeTokenOwnership',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'bool',
+        name: 'add',
+        type: 'bool',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'amount',
+        type: 'uint256',
+      },
+    ],
+    name: 'UpdateBalanceLocked',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'bool',
+        name: 'add',
+        type: 'bool',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'amount',
+        type: 'uint256',
+      },
+    ],
+    name: 'UpdateBalancePooled',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'newBalance',
+        type: 'uint256',
+      },
+    ],
+    name: 'UpdateBalances',
+    type: 'event',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'cancelResubscribe',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'cancelTransferMeTokenOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'oldOwner',
+        type: 'address',
+      },
+    ],
+    name: 'claimMeTokenOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'finishResubscribe',
+    outputs: [
+      {
+        components: [
+          {
+            internalType: 'address',
+            name: 'owner',
+            type: 'address',
+          },
+          {
+            internalType: 'uint256',
+            name: 'hubId',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'balancePooled',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'balanceLocked',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'startTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endCooldown',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'targetHubId',
+            type: 'uint256',
+          },
+          {
+            internalType: 'address',
+            name: 'migration',
+            type: 'address',
+          },
+        ],
+        internalType: 'struct MeTokenInfo',
+        name: '',
+        type: 'tuple',
+      },
+    ],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+    ],
+    name: 'getMeTokenInfo',
+    outputs: [
+      {
+        components: [
+          {
+            internalType: 'address',
+            name: 'owner',
+            type: 'address',
+          },
+          {
+            internalType: 'uint256',
+            name: 'hubId',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'balancePooled',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'balanceLocked',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'startTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endTime',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'endCooldown',
+            type: 'uint256',
+          },
+          {
+            internalType: 'uint256',
+            name: 'targetHubId',
+            type: 'uint256',
+          },
+          {
+            internalType: 'address',
+            name: 'migration',
+            type: 'address',
+          },
+        ],
+        internalType: 'struct MeTokenInfo',
+        name: '',
+        type: 'tuple',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+    ],
+    name: 'getOwnerMeToken',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'oldOwner',
+        type: 'address',
+      },
+    ],
+    name: 'getPendingOwner',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'targetHubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'address',
+        name: 'migration',
+        type: 'address',
+      },
+      {
+        internalType: 'bytes',
+        name: 'encodedMigrationArgs',
+        type: 'bytes',
+      },
+    ],
+    name: 'initResubscribe',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+    ],
+    name: 'isOwner',
+    outputs: [
+      {
+        internalType: 'bool',
+        name: '',
+        type: 'bool',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'meTokenCooldown',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'meTokenDuration',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'meTokenWarmup',
+    outputs: [
+      {
+        internalType: 'uint256',
+        name: '',
+        type: 'uint256',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'cooldown',
+        type: 'uint256',
+      },
+    ],
+    name: 'setMeTokenCooldown',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'duration',
+        type: 'uint256',
+      },
+    ],
+    name: 'setMeTokenDuration',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'uint256',
+        name: 'warmup',
+        type: 'uint256',
+      },
+    ],
+    name: 'setMeTokenWarmup',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'string',
+        name: 'name',
+        type: 'string',
+      },
+      {
+        internalType: 'string',
+        name: 'symbol',
+        type: 'string',
+      },
+      {
+        internalType: 'uint256',
+        name: 'hubId',
+        type: 'uint256',
+      },
+      {
+        internalType: 'uint256',
+        name: 'assetsDeposited',
+        type: 'uint256',
+      },
+    ],
+    name: 'subscribe',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
+    ],
+    name: 'transferMeTokenOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'meToken',
+        type: 'address',
+      },
+      {
+        internalType: 'uint256',
+        name: 'newBalance',
+        type: 'uint256',
+      },
+    ],
+    name: 'updateBalances',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'deactivateController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'diamondController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'durationsController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'feesController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'meTokenRegistryController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'registerController',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setDeactivateController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setDiamondController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setDurationsController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setFeesController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setMeTokenRegistryController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'newController',
+        type: 'address',
+      },
+    ],
+    name: 'setRegisterController',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      {
+        internalType: 'address',
+        name: 'forwarder',
+        type: 'address',
+      },
+    ],
+    name: 'setTrustedForwarder',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'trustedForwarder',
+    outputs: [
+      {
+        internalType: 'address',
+        name: '',
+        type: 'address',
+      },
+    ],
+    stateMutability: 'view',
+    type: 'function',
   },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setDurationsController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setFeesController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setMeTokenRegistryController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "newController",
-  "type": "address"
-  }
-  ],
-  "name": "setRegisterController",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [
-  {
-  "internalType": "address",
-  "name": "forwarder",
-  "type": "address"
-  }
-  ],
-  "name": "setTrustedForwarder",
-  "outputs": [],
-  "stateMutability": "nonpayable",
-  "type": "function"
-  },
-  {
-  "inputs": [],
-  "name": "trustedForwarder",
-  "outputs": [
-  {
-  "internalType": "address",
-  "name": "",
-  "type": "address"
-  }
-  ],
-  "stateMutability": "view",
-  "type": "function"
-  }
 ]
-
 
 // DAI ADDRESS
 export const DAI = {
@@ -3250,7 +3232,6 @@ export const ERC20_ABI = [
     type: 'function',
   },
 ]
-
 
 // ETH MAINNET THIRDWEB RPC
 export const MAINNET_RPC = process.env.NEXT_PUBLIC_ALCHEMY_API_KEY || 'https://ethereum.rpc.thirdweb.com'
